app/engine/pyproject.toml
[project]
name = "ii-engine"
version = "0.1.0"
description = "Orchestration engine for IntrusionInspector"
requires-python = ">=3.12"
dependencies = [
    "ii-settings",
    "ii-logger",
    "ii-errors",
    "ii-utils",
    "ii-collectors",
    "ii-analyzers",
    "ii-reporters",
    "ii-evidence",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/engine"]
